Create Lab 4-Twenty One Game
This game creates a value for a player and a computer. It offers the player to "hit" (add to their value) or "stand" (keep that value). If the player has a value over 21, they lose. They also lose if their value is under 21 and the computer has a higher value. Their final value needs to be 21 or lower AND be higher than the computer's.